+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 # Changelog
 ## Current
 ### Features
+* Refactor rights to be able to check inheritance and add rights on model
+* Add rating form route and model
+* Add push notifications
+* User can delete his own account
 ### Fix
+* rides and ratings routes tests
 ## Version 1.4.3
 ### Features
 * Add /campuses/{id}/users route